+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Last Changes
 
+- [#5](https://github.com/LaxarJS/grunt-init-laxar-application/issues/5): Added json patch info to require configuration.
 - [#4](https://github.com/LaxarJS/grunt-init-laxar-application/issues/4): jshintrc: disabled enforcement of dot notation for object property access.
 - [#3](https://github.com/LaxarJS/grunt-init-laxar-application/issues/3): Added RequireJS configuration necessary for new JSON validator.
 - [#1](https://github.com/LaxarJS/grunt-init-laxar-application/issues/1): Added `debug.html` to template.